+++ src/apps/backend/modules/account/internal/account-writer.ts
@@ -1,12 +1,12 @@
-import { ConfigService } from '../../config';
+import { PasswordResetTokenService } from '../../password-reset-token';
+import PasswordResetTokenUtil from '../../password-reset-token/internal/password-reset-token-util';
 import {
-  Account, AccountBadRequestError, CreateAccountParams, PasswordResetToken,
+  Account, AccountBadRequestError, CreateAccountParams,
 } from '../types';
 
 import AccountReader from './account-reader';
 import AccountUtil from './account-util';
 import AccountRepository from './store/account-repository';
-import PasswordResetTokenRepository from './store/password-reset-token-repository';
 
 export default class AccountWriter {
   public static async createAccount(
@@ -28,32 +28,12 @@ export default class AccountWriter {
     return AccountUtil.convertAccountDBToAccount(accDb);
   }
 
-  public static async createPasswordResetToken(
-    accountId: string,
-    token: string,
-  ): Promise<PasswordResetToken> {
-    const defaultTokenExpireTimeInSeconds = ConfigService.getValue<string>(
-      'passwordResetToken.expiresInSeconds',
-    );
-    const tokenHash = await AccountUtil.hashPasswordResetToken(token);
-
-    const passwordResetTokenDB = await PasswordResetTokenRepository.create({
-      account: accountId,
-      expiresAt: AccountUtil.getTokenExpiresAt(defaultTokenExpireTimeInSeconds),
-      token: tokenHash,
-    });
-
-    return AccountUtil.convertPasswordResetTokenDBToPasswordResetToken(
-      passwordResetTokenDB,
-    );
-  }
-
   public static async resetPassword(
     accountId: string,
     newPassword: string,
     token: string,
   ): Promise<Account> {
-    const passwordResetToken = await AccountReader.getPasswordResetTokenByAccountId(accountId);
+    const passwordResetToken = await PasswordResetTokenService.getPasswordResetTokenByAccountId(accountId);
 
     if (passwordResetToken.isExpired) {
       throw new AccountBadRequestError(
@@ -67,7 +47,7 @@ export default class AccountWriter {
       );
     }
 
-    const isTokenValid = await AccountUtil.comparePasswordOrResetToken(
+    const isTokenValid = await PasswordResetTokenUtil.comparePasswordResetToken(
       token,
       passwordResetToken.token,
     );
@@ -86,21 +66,10 @@ export default class AccountWriter {
       },
     );
 
-    await AccountWriter.setPasswordResetTokenAsUsed(
+    await PasswordResetTokenService.setPasswordResetTokenAsUsedById(
       passwordResetToken.id,
     );
 
     return AccountUtil.convertAccountDBToAccount(dbAccount);
-  }
-
-  private static async setPasswordResetTokenAsUsed(
-    passwordResetTokenId: string,
-  ): Promise<PasswordResetToken> {
-    return PasswordResetTokenRepository.findByIdAndUpdate(
-      passwordResetTokenId,
-      {
-        used: true,
-      },
-    );
   }
 }
