+++ src/apps/backend/modules/account/rest-api/account-controller.ts
@@ -2,7 +2,7 @@ import { applicationController, Request, Response } from '../../application';
 import { HttpStatusCodes } from '../../http';
 import AccountService from '../account-service';
 import {
-  CreateAccountParams, GetAccountParams, CreatePasswordResetTokenParams, ResetPasswordParams,
+  CreateAccountParams, GetAccountParams, ResetPasswordParams,
 } from '../types';
 
 import { serializeAccountAsJSON } from './account-serializer';
@@ -30,16 +30,6 @@ export class AccountController {
       const accountJSON = serializeAccountAsJSON(account);
 
       res.status(HttpStatusCodes.OK).send(accountJSON);
-    },
-  );
-
-  createPasswordResetToken = applicationController(
-    async (req: Request<CreatePasswordResetTokenParams>, res: Response) => {
-      await AccountService.createPasswordResetToken({
-        username: req.body.username,
-      });
-
-      res.status(HttpStatusCodes.CREATED).send();
     },
   );
