+++ src/apps/backend/modules/account/rest-api/account-router.ts
@@ -10,8 +10,6 @@ export default class AccountRouter extends ApplicationRouter {
 
     router.post('/', ctrl.createAccount);
 
-    router.post('/password-reset-token', ctrl.createPasswordResetToken);
-
     router.patch('/:accountId/reset-password', ctrl.resetPassword);
 
     router.use(accessAuthMiddleware);
