+++ src/apps/backend/modules/password-reset-token/index.ts
@@ -1,0 +1,3 @@
+export { default as PasswordResetTokenServer } from './rest-api/password-reset-token-server';
+export { default as PasswordResetTokenService } from './password-reset-token-service';
+export * from './types';
