+++ src/apps/backend/modules/password-reset-token/rest-api/password-reset-token-router.ts
@@ -1,0 +1,12 @@
+import { ApplicationRouter } from '../../application';
+import { PasswordResetTokenController } from './password-reset-token-controller';
+
+
+export default class PasswordResetTokenRouter extends ApplicationRouter {
+  configure(): void {
+    const { router } = this;
+    const ctrl = new PasswordResetTokenController();
+
+    router.post('/password-reset-token', ctrl.createPasswordResetToken);
+  }
+}
